Bug#20573701 DROP DATABASE ASSERT ON DEBUG WHEN OTHER FILES PRESENT IN DB FOLDER.
Analysis: In the 'DROP DATABASE' execution, the condition to check and
throw error (ER_DB_DROP_RMDIR) if there are extra files in the data
directory was wrongly placed in if(binlog is opened) block.
Hence this error case was not caught when binlog is disabled.

Fix: Moved the checking code to the proper place which will be executed
in all the cases (binlog enabled/disabled)