Bug#20856729: QUERY REWRITE: WRONG IFDEF SYMBOL IN SERVICE_PARSER.H
The #ifdef'd symbol MYSQL_DYNAMIC_PLUGIN was misspelled, causing the code
using a struct mysql_parser_service_st to disappear. When this was fixed, it
became evident that this code hadn't kept up with the development of the
service API, the function mysql_parser_get_statement_type() was missing. The
naming was also inconsistent, so it has now been changed to this:

- The function pointers inside the struct have names starting with mysql_.
- The free functions pointed to have names starting with mysql_parser_.

The public API changes, but it was incorrect before.