Fix Bug#20618309 ASSERT SLOT1->PAGE_LEVEL == SLOT2->PAGE_LEVEL, BTR_ESTIMATE_N_ROWS_IN_RANGE()
Relax a too strict assert. If the tree is changed between both dives for
the left boundary and the right boundary, then our markers (slot1 and
slot2) could end up on different levels in the tree.

If we detect that this has happened - then retry a few times and if
still unsuccessful then return an arbitrary number for an estimate.

Reviewed-by:	Marko Mäkelä <marko.makela@oracle.com>
Reviewed-by:	Jimmy Yang <jimmy.yang@oracle.com>
RB:		8570