Bug#20902791 MYSQLDUMP DUMPS SYS_SCHEMA
Do not dump the sys schema when using mysqldump --all-databases.

Note: It is still possible to dump the sys schema individually (mysqldump --databases sys), as it may still be something that a user wants to do (given the sys.sys_config persistent table could be recovered for instance if the user wanted to). Further, the sys schema is not ignored for LOCK TABLES as with the performance_schema and INFORMATION_SCHEMA, as the sys schema stores "regular" (InnoDB) objects, that option is still feasible.

Reviewed-by: Bharathy X Satish <bharathy.x.satish@oracle.com>
Reviewed-by: Georgi Kodinov <georgi.kodinov@oracle.com>
RB: 8677