Bug#20927239: MY_TIMER-T UNIT TEST DOES NOT WORK WITH MERGE_UNITTESTS=0
The problem was that the my_timer-t unit test did not work
properly if it was compiled separately. This was because the
unit test code was wrapped in #if HAVE_MY_TIMER. This symbol
is defined in my_config.h which had not been included at that
point. The same problem existed in my_timer.h.

This patch fixes the problem by removing HAVE_MY_TIMER.
This symbol is always true for all supported platforms anyway.
Removing this makes the unit test work as intended and
simplifies code.