BUG#20889900: UNITTESTS SHOULD START THE SERVER WITH APPROPRIATE OPTIONS
Problem: Unittests which start a server will log errors to syslog, emit a
number of warnings, and also complain about not finding the error messages
in the default location. The latter means that unittests cannot be used to
properly test error handling.

Solution: Add necessary startup options passed to the server
when running unittests.

Remove some unneeded includes in debug_sync-t.cc which caused a build
break on WIN with MERGE_UITTESTS=0