WL#6940 Server version token and check
Version Tokens can help a system of clients and servers to
synchronize with each other. It is implemented using a
plugin named version_token. Tokens can be set on both global
level and session level and the status check is done by
comparing session level tokens with the global tokens before
every query.

UDFs are provided to manage the global tokens.