Bug#20947871 INDEX SCAN COST IN TEST_IF_CHEAPER_ORDERING() DOES NOT USE COST CONSTANTS
The cost calculation for the cost estimate of doing index scan in
test_if_cheaper_ordering() does not use the configurable cost constant
for the io cost of accessing a block from the table. Instead it assumes
that all accesses to a block cost 1:

  const double index_scan_time= select_limit / rec_per_key *
     min<double>(rec_per_key, table_scan_time.total_cost());

The the last rec_per_key in this formula corresponds to the number of
page reads that is needed to read from the base table.

This can cause unintended changes to query plans if the cost constant
for io block read is changed. The fix for this issue is to use the
configurable cost constant.