Fix a possible crash in AutoTest when an ordered scan encounter error 4008, scan timeout. One such testcase is 'testScan -n ScanRead4880'
In case scan timeouts are encountered, the test attempts to reduce the
load by changing to unordered scans. This is done by setting the
'const NdbDictionary::Index * pIdx' argument to NULL inside
HugoTransactions::scanReadRecords.

As pIdx->getName() is later called when creating a NdbIndexScanOperation
we will crash.

The patch below fixes this issue by calling NdbScanOperation() instead
if 'pIdx == NULL'