Bug#20782580: NDB REDO OVERLOAD: ABORT QUEUING REQUESTS IF OUT OF LOG SPACE
If redo log writes are requested when the redo log is full, the
writes are queued instead of being rejected. These requests have
to wait in queue until redo space is freed. However, redo space
is freed only when the next LCP completes, which can take a long
time. This results in long waits and timeouts.

Modified redo log queuing to abort requests which are received
when the redo log is full. The requests are aborted with the
error code 410 'REDO log files overloaded (decrease
TimeBetweenLocalCheckpoints or increase NoOfFragmentLogFiles)'.