Bug#20842030 ST_BUFFER CAN BE CALLED WITH PARAMETERS TO CONSUME INORDINATE AMOUNT OF MEMORY
Define a session variable 'max_points_in_geometry' which is allowed in range
[3, 64*1024*1024] as a per-session customizable limit for the maximum number of
points in a geometry. The upper bound for the range is got
using max_allowed_packet/16, where 16 is one point's storage size.

In ST_Buffer strategy specification, reject 'points_per_circle' arguments
greater than this limit, in order for users to limit the memory
and CPU consumption in buffer calculation.