Bug #20294225 - INVALID MEMORY ACCESS
DESCRIPTION
===========
/client/mysql_config_editor.cc:

In encrypt_and_write_file(), my_write() is called even if 'enc_len' is
4096(which it is allowd to be). In such a case it will be accessed at
offset 4099 which will result in SEGFAULT!

ANALYSIS
========
The crash can be avoided if the encrypted buffer size check is put in
place (by calling my_aes_get_size()) even before doing encryption.

FIX
===
Do the above mentioned check in encrypt_and_write_file(). Since to do
this we already call my_aes_get_size() for getting length of
to-be-encrypted-buffer, this function call is removed from
encrypt_buffer(). This length is passed to encrypt_buffer() as an
additional argument.