BUG#19988193: ASSERTION `(*TABLES)->REGINFO.LOCK_TYPE >= TL_READ' FAILED IN LOCK_EXTERNAL BUG#21198646: ASSERTION FAILED: (*TABLES)->REGINFO.LOCK_TYPE >= TL_READ FILE LOCK.CC, LINE 356
This patch addresses two related issues: Calling a procedure which
creates a view from a trigger (BUG#19988193), and creating a function
calling a procedure doing RENAME TABLE (BUG#21198646), could both, in
certain circumstances, trigger an assert.

Root cause was that prelocking of tables with lock_type==TL_IGNORE is
not supported, and so triggers an assert. TL_IGNORE is only used for
source tables in CREATE VIEW statements and the table of a RENAME TABLE
statement. It is very unusual for these statements to be part of
prelocking analysis, as both are implicit commit statements which are
not permitted in triggers and stored functions/procedures. But as the test
cases show; it is possible to have such statements contribute to the
prelocking set, but in both cases the statement is "meaningless", in
the sense that it will trigger an error during execution.

Fix: In mysql_make_view(), avoid adding the backing tables to view_ref
if view_ref->prelocking_placeholder==true and lock_type==TL_IGNORE. In
sp_head::add_used_tables_to_table_list() skip SP_TABLES which have
lock_type=TL_IGNORE.

Test: New test cases added to tablelock.test