Bug#20748537 INNODB: FAILING ASSERTION: NODE->PCUR->REL_POS == BTR_PCUR_ON
The intention of sql_buffer_results is to store the result into a
tmp table so that locks are freed early (don't need to wait for client's
ok).
Multi-table UPDATE/DELETE don't allow sql_buffer_results as a keyword in the
query (only SELECT does).
But buffering can still be requested by setting @@sql_buffer_results; that
variable is documented to affect only SELECT:
"If set to 1, sql_buffer_result forces results from SELECT statements
to be put into temporary tables". I can't imagine what this variable
would mean for multi-table UPDATE/DELETE. Multi-table DELETE	already
turns it off.
We here do the same for multi-table UPDATE. Letting it on caused
a crash in InnoDB (cursor in an invalid state).