Bug#20379981 ST_CONTAINS(GEOMETRYCOLLECTION(POLY,POLY), LINESTRING) RETURNS WRONG RESULT
Issue:

The within/contains algorithm for geometry collections can't work correctly
if in "M within N" check, N is some polygons that touch (do not overlap)
or some linestrings that touch (do not overlap), and M is a
[multi]linestring/[multi]polygon that is entirely in multiple components of N.

Fix:

When merging components of a geometry collection(GC), do union to any
pair of components that intersect, except if one is linear(L) and the other is
areal(A), and they intersect only finite number of points. That is, if L either
crosses A or L overlaps the outer ring of A, replace the L and A
with ((L - A) U A).

In the result of such a merge, the only intersecting components of a GC are
[multi]linestrings intersecting [multi]polygons at finite number of points,
all other kinds of intersecting components are merged.

And such a fix also fixes this issue:

ST_CONTAINS(GEOMETRYCOLLECTION(LINESTRING1,LINESTRING2), LINESTRING3)

where LINESTRING1 touches LINESTRING2, and LINESTRING3 is not within
LINESTRING1 or LINESTRING2 but is is entirely within
(LINESTRING1 union LINESTRING2).

and it also fixes this issue:

ST_CONTAINS(GEOMETRYCOLLECTION(POLY1,POLY2), POLY3)

where POLY1 touches POLY2, and POLY3 is not within POLY1 or POLY2 but is
entirely within (POLY1 union POLY2).