BUG#20977779 CANNOT IMPORT TABLES CONTAINING PREFIX INDEXES
Problem :
---------
While importing table, we are getting the index columns
from cfg by name and hence if a column is repeated we get
the first reference always. If the first and second reference
has different prefix length, the validation fails.

Solution :
----------
Check all index column in dictionary and cfg one to one. The
fix expects cfg and dictionary to match exactly.

Reviewed-by: Kevin Lewis <kevin.lewis@oracle.com>
Reviewed-by: Marko Makela <marko.makela@oracle.com>

RB: 8774