bug#18411034: Assertion gci_ops && (gci == gci_ops->m_gci) failed
The assertion fails due to the event queue not being cleared after the
subscriptions are dropped and recreated due to a cluster initial
restart, and continue using the same ndb object. In this case, the
event queue contains epochs belonging to two incarnations of the
cluster (higher epochs from first incarnation followed by lower epochs
from the next one), causing the assertion failure.