Bug#19316063: MAKE MTS WORK WITH RELAY_LOG_RECOVERY=1 WHEN GTID IS ENABLED
Analysis:
========
When GTID is enabled, MTS don't need to care about gaps.

If the auto position replication protocol is enabled, it
will make dump thread handle all of the gaps on master side.

If the auto position replication protocol is not enabled,
the dump thread will start replicating from the first gap
and the GTID auto skip feature will make sure that the
already applied transactions will not be applied twice.

Fix:
===
During relay log recovery process check if GTID mode is
enabled or not. If enabled reset the MTS recovery process
so that GTID protocol can fill the MTS gaps.