BUG#20768847 ASSERTION 0 IN HA_INNOBASE::COMMIT_INPLACE_ALTER_TABLE
Problem :
---------
We are incorrectly dropping an index in parent table with foreign key
dependency. While checking whether it is ok to drop an index, we check
only the first FK reference to and from other tables. For multiple
foreign key references the check is insufficient.

Solution :
----------
Loop and check for all foreign key references to and from the table to
check if it is safe to drop an index.

Reviewed-by: Marko Makela <Marko.Makela@oracle.com>

RB: 8734