Bug#20379981 ST_CONTAINS(GEOMETRYCOLLECTION(POLY,POLY), LINESTRING) RETURNS WRONG RESULT
Follow-up fix: the previous patch for this bug turned out to be the last straw
that made the build on Solaris fail because virtual memory is exhausted when
building item_geofunc_buffer.cc. The root cause for such failure is not the
changes made in previous patch for this bug,
but how we made use of boost::geometry::buffer when implementing ST_Buffer---
we used boost::variant to hold 5 types of buffer strategies to call different
versions of boost::geometry::buffer according to the strategy combinations
without having to enumerate all the combinations.
Such usage takes much memory to compile the source code.

Now no longer use boost::variant, and dispatch boost::geometry::buffer calls
according to the strategies specified in an efficient way.

And such a change also fixed an issue:

Before this change, if a type of strategy, e.g. end_strategy,
isn't specified using st_buffer_strategy() call, and it's useless for the type
of input argument, the strategies given to boost::geometry::buffer() is not
what we defined for ST_Buffer.

Taking below query as example:

	ST_Buffer(ST_GeomFromText('Point(0 0)'))

here the end strategy and join strategy applied to boost::geometry::buffer
is end_round and join_round with a BG internally defined 'points_per_circle'
value rather than the default value(32) we defined for ST_Buffer.

Although logically it doesn't make sense for the end and join strategies to be
used in this case, boost::geometry::buffer() does need all strategy arguments
no matter what type of input geometry is given, and the end/join strategies do
have effect on the result here, this is determined by how bg::buffer is
implemented.

Now always use default strategies as defined for the ST_Buffer function if not
specified explicitly, even if the type of strategy isn't allowed for the type
of input geometry according to our definition to ST_Buffer.