Bug #20376498: MAX_ALLOWED_PACKET ERROR DESTROYS ORIGINAL               DATA
Issue:
-----
When a string exceeds the max_allowed_packet size, a null
value is returned and a warning issued. In an update
statement, this can result replacing of existing value with
a null value.

Essentially, resulting in a loss of existing data.

SOLUTION:
---------
An error is thrown if all the below conditions are
satisfied:
1) max_allowed_packet is exceeded.
2) Operation involves a table write.
3) Strict mode is set.