Bug#20835095 CRASH AT CREATE_REF_FOR_KEY IN SQL/SQL_SELECT.CC
Description:
------------
After bug#19695490 was fixed, a special case missed, that leads to this bug.

For a VARCHAR type, if MySQL uses it as a key of temp table, MySQL will add
HA_KEY_BLOB_LENGTH to the store length in order to calculate the key length
in init_from_field(), so we add HA_KEY_BLOB_LENGTH to the length of a
possible index to prevent the optimizer to use to materialized lookup.
If we don't do this, MySQL would triggers an assertion, which was added in
bug#19695490.

Fix:
----
Add HA_KEY_BLOB_LENGTH to 'lookup_index_length' in order to get the right
index length in semijoin_types_allow_materialization() when the field is
VARCHAR, BLOB or GEOMETRY.

Test case added.